+++ src/modules/user/dto/CreateUserRoleDto.ts
@@ -1,0 +1,6 @@
+import { UserRole } from "../../../common/enums/UserRole";
+
+export interface CreateUserRoleDto {
+    userId: string;
+    roles: UserRole[]
+}
